🚸 fix project detail loading ui #28
